Create list of shuffled integers
Added the createRandomList class which creates a list of 1,000,000
integers then shuffles them and saves them to a file.
oneSixThree.java will then read this file, sort it using the bitwise
operators, then print it to the screen. I may come back to this exercise
later to improve it, but for now I'm considering this complete.